huella de carbono calculado y agente sectorial creo que hecho

+++ src/main/java/utn/ddsG8/impacto_ambiental/movilidad/transportes/Transporte.java
@@ -1,8 +1,8 @@
 package utn.ddsG8.impacto_ambiental.movilidad.transportes;
 
-import utn.ddsG8.impacto_ambiental.estructura.CalcularHC;
+import utn.ddsG8.impacto_ambiental.calculos.CalcularHC;
 
 public interface Transporte {
 
-    float calcularHC(CalcularHC calculador,float distancia);
+    double calcularHC(CalcularHC calculador,double distancia);
 }
