+++ src/main/java/utn/ddsG8/impacto_ambiental/movilidad/transportes/publico/TransportePublico.java
@@ -1,8 +1,7 @@
 package utn.ddsG8.impacto_ambiental.movilidad.transportes.publico;
 
-import utn.ddsG8.impacto_ambiental.estructura.CalcularHC;
+import utn.ddsG8.impacto_ambiental.calculos.CalcularHC;
 import utn.ddsG8.impacto_ambiental.movilidad.transportes.Transporte;
-import utn.ddsG8.impacto_ambiental.movilidad.transportes.publico.Parada;
 
 import java.util.List;
 
@@ -11,7 +10,7 @@ public abstract class TransportePublico implements Transporte {
     protected String nombreFE;
     protected List<Parada> paradas;
 
-    public void agregarParada(Parada nueva, float distProxima, float distAnterior, int i) {
+    public void agregarParada(Parada nueva, double distProxima, double distAnterior, int i) {
         // IMPORTANTE: Esta implementado como que el administrador pone 0 para la primera parada, no 1
         // Se puede cambiar con un simple -1. Decidir que queda
         paradas.add(i, nueva);
@@ -34,9 +33,9 @@ public abstract class TransportePublico implements Transporte {
         return paradaIncial.distanciaAParada(paradaFinal,this);
     }*/
 
-    public float calcularHC(CalcularHC calculador, float distancia) {
-        float fe;
-        fe = (float) calculador.buscarFactorEmision(this.nombreFE,"");
+    public double calcularHC(CalcularHC calculador, double distancia) {
+        double fe;
+        fe = (double) calculador.buscarFactorEmision(this.nombreFE,"");
         if (fe != -1){
             return fe * distancia;
         }
