+++ src/main/java/utn/ddsG8/impacto_ambiental/services/distancia/Distancia.java
@@ -1,10 +1,10 @@
 package utn.ddsG8.impacto_ambiental.services.distancia;
 
 public class Distancia {
-    public float valor;
+    public double valor;
     public String unidad;
 
-    public Distancia(float valor, String unidad) {
+    public Distancia(double valor, String unidad) {
         this.valor = valor;
         this.unidad = unidad;
     }
